Improved game deletion errors

+++ Spring/src/main/java/pl/ttsw/GameRev/controller/GameController.java
@@ -6,19 +6,27 @@ import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import pl.ttsw.GameRev.dto.GameDTO;
 import pl.ttsw.GameRev.filter.GameFilter;
+import pl.ttsw.GameRev.model.Forum;
+import pl.ttsw.GameRev.model.ForumRequest;
+import pl.ttsw.GameRev.repository.ForumRepository;
+import pl.ttsw.GameRev.repository.ForumRequestRepository;
 import pl.ttsw.GameRev.service.GameService;
 import java.io.IOException;
+import java.util.List;
 
 @RestController
 @RequestMapping("/games")
 @RequiredArgsConstructor
 public class GameController {
     private final GameService gameService;
+    private final ForumRepository forumRepository;
+    private final ForumRequestRepository forumRequestRepository;
 
     @GetMapping
     public ResponseEntity<Page<GameDTO>> getAllGames(GameFilter gameFilter, Pageable pageable) {
@@ -63,6 +71,16 @@ public class GameController {
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<?> deleteGame(@PathVariable Long id) {
+        List<Forum> forums = forumRepository.findByGameId(id);
+        if (!forums.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("Game is used in forums");
+        }
+
+        List<ForumRequest> forumRequests = forumRequestRepository.findAllByGameId(id);
+        if (!forumRequests.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("Game is used in forum requests");
+        }
+
         boolean gotRemoved = gameService.deleteGame(id);
         if (!gotRemoved) {
             return ResponseEntity.notFound().build();
